Migrate to DDE using Shared Makefiles
Move from native GCC to DDE, using the new Shared Makefile format.

The compilation process requires the CompileDate utility to be present
in the Library or somewhere else on Run$Path in order to be able to
create the necessary h.compiledate file. If not available, then this
file needs to define INFO_BOX_DATE along the following lines:

    #define INFO_BOX_DATE "2 Apr 2025"